+++ app/src/main/java/com/muso/music/MainActivity.kt
@@ -176,7 +176,6 @@ import com.muso.music.constants.SYSTEM_DEFAULT
 import java.util.Locale
 import kotlinx.coroutines.runBlocking
 import com.muso.music.utils.dataStore
-import com.muso.music.utils.edit
 import com.muso.music.constants.LastArtworkBackgroundColorKey
 import com.muso.music.constants.ArtworkBackgroundKey
 import androidx.datastore.preferences.core.edit