+++ app/src/main/java/com/muso/music/ui/player/Player.kt
@@ -700,7 +700,7 @@ fun BottomSheetPlayer(
                     )
                 }
 
-                SliderStyle.SQUIGGLY -> {
+                sliderStyle == SliderStyle.SQUIGGLY -> {
                     SquigglySlider(
                         value = (sliderPosition ?: position).toFloat(),
                         valueRange = 0f..(if (duration == C.TIME_UNSET) 0f else duration.toFloat()),